Renamed Client to Validator (#428)
