+++ validator/BUILD.bazel
@@ -5,13 +5,13 @@ load("@io_bazel_rules_docker//container:container.bzl", "container_push")
 go_library(
     name = "go_default_library",
     srcs = ["main.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator",
+    visibility = ["//validator:__subpackages__"],
     deps = [
-        "//client/node:go_default_library",
-        "//client/types:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
+        "//validator/node:go_default_library",
+        "//validator/types:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
         "@com_github_x_cray_logrus_prefixed_formatter//:go_default_library",
@@ -23,15 +23,15 @@ go_image(
     srcs = ["main.go"],
     goarch = "amd64",
     goos = "linux",
-    importpath = "github.com/prysmaticlabs/prysm/client",
+    importpath = "github.com/prysmaticlabs/prysm/validator",
     static = "on",
     tags = ["manual"],
     visibility = ["//visibility:private"],
     deps = [
-        "//client/node:go_default_library",
-        "//client/types:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
+        "//validator/node:go_default_library",
+        "//validator/types:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
         "@com_github_x_cray_logrus_prefixed_formatter//:go_default_library",
@@ -43,13 +43,13 @@ container_push(
     format = "Docker",
     image = ":image",
     registry = "gcr.io",
-    repository = "prysmaticlabs/prysm/client",
+    repository = "prysmaticlabs/prysm/validator",
     tag = "latest",
     tags = ["manual"],
 )
 
 go_binary(
-    name = "client",
+    name = "validator",
     embed = [":go_default_library"],
-    visibility = ["//client:__subpackages__"],
+    visibility = ["//validator:__subpackages__"],
 )
