+++ validator/attester/BUILD.bazel
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["service.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/attester",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/attester",
+    visibility = ["//validator:__subpackages__"],
     deps = [
-        "//client/types:go_default_library",
+        "//validator/types:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
