+++ validator/beacon/BUILD.bazel
@@ -3,11 +3,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["service.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/beacon",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/beacon",
+    visibility = ["//validator:__subpackages__"],
     deps = [
-        "//client/types:go_default_library",
         "//proto/beacon/rpc/v1:go_default_library",
+        "//validator/types:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_bazel_rules_go//proto/wkt:empty_go_proto",
@@ -20,10 +20,10 @@ go_test(
     srcs = ["service_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//client/internal:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/beacon/rpc/v1:go_default_library",
         "//shared/testutil:go_default_library",
+        "//validator/internal:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
