+++ validator/internal/BUILD.bazel
@@ -4,8 +4,8 @@ go_library(
     name = "go_default_library",
     testonly = True,
     srcs = ["beacon_service_mock.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/internal",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/internal",
+    visibility = ["//validator:__subpackages__"],
     deps = [
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/beacon/rpc/v1:go_default_library",
