+++ validator/internal/beacon_service_mock.go
