+++ validator/node/BUILD.bazel
@@ -3,20 +3,20 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["node.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/node",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/node",
+    visibility = ["//validator:__subpackages__"],
     deps = [
-        "//client/attester:go_default_library",
-        "//client/beacon:go_default_library",
-        "//client/proposer:go_default_library",
-        "//client/rpcclient:go_default_library",
-        "//client/txpool:go_default_library",
-        "//client/types:go_default_library",
         "//shared:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/database:go_default_library",
         "//shared/debug:go_default_library",
         "//shared/p2p:go_default_library",
+        "//validator/attester:go_default_library",
+        "//validator/beacon:go_default_library",
+        "//validator/proposer:go_default_library",
+        "//validator/rpcclient:go_default_library",
+        "//validator/txpool:go_default_library",
+        "//validator/types:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
     ],
