+++ validator/params/BUILD.bazel
@@ -3,8 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["config.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/params",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/params",
+    visibility = ["//validator:__subpackages__"],
 )
 
 go_test(
