+++ validator/proposer/BUILD.bazel
@@ -3,23 +3,23 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["service.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/rpcclient",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/proposer",
+    visibility = ["//validator:__subpackages__"],
     deps = [
-        "//proto/beacon/rpc/v1:go_default_library",
+        "//validator/types:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//credentials:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
+    size = "small",
     srcs = ["service_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//proto/beacon/rpc/v1:go_default_library",
         "//shared/testutil:go_default_library",
+        "@com_github_golang_mock//gomock:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
 )
