+++ validator/proposer/service.go
@@ -5,7 +5,7 @@ package proposer
 import (
 	"context"
 
-	"github.com/prysmaticlabs/prysm/client/types"
+	"github.com/prysmaticlabs/prysm/validator/types"
 	"github.com/sirupsen/logrus"
 )
 
@@ -17,11 +17,11 @@ var log = logrus.WithField("prefix", "proposer")
 type Proposer struct {
 	ctx           context.Context
 	cancel        context.CancelFunc
-	beaconService types.BeaconClient
+	beaconService types.BeaconValidator
 }
 
 // NewProposer creates a new attester instance.
-func NewProposer(ctx context.Context, beaconService types.BeaconClient) *Proposer {
+func NewProposer(ctx context.Context, beaconService types.BeaconValidator) *Proposer {
 	ctx, cancel := context.WithCancel(ctx)
 	return &Proposer{
 		ctx:           ctx,
