+++ validator/types/BUILD.bazel
@@ -8,12 +8,12 @@ go_library(
         "interfaces.go",
         "shard.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/client/types",
-    visibility = ["//client:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/types",
+    visibility = ["//validator:__subpackages__"],
     deps = [
-        "//client/params:go_default_library",
         "//proto/beacon/rpc/v1:go_default_library",
         "//shared:go_default_library",
+        "//validator/params:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//core/types:go_default_library",
         "@com_github_ethereum_go_ethereum//crypto/sha3:go_default_library",
